% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-accessors.R
\name{DataTableSchema}
\alias{DataTableSchema}
\title{Data table schema object}
\description{
An object that describes the schema of a data table, like a worksheet's
summary data or underlying data, or a logical table from a data source.
}
\section{Fields}{

\describe{
\item{\code{id}}{Character vector indicating the ID of the underlying data table or
logical table. (Not present for summary data.)}

\item{\code{caption}}{Character vector with a human-readable description of the
underlying data table or logical table. (Not present for summary data.)}

\item{\code{name}}{Character vector with either \code{"Underlying Data Table"} or
\code{"Summary Data Table"} (yes, literally those strings; not sure why this
field is called \code{name} but it comes from the Tableau Extension API).}

\item{\code{columns}}{A data frame that describes the columns in this table. Each
data frame row describes a column in the data table. The data frame
contains these columns:
\itemize{
\item \code{dataType} - character - \code{"float"}, \code{"integer"}, \code{"string"}, \code{"boolean"}, \code{"date"}, or \code{"datetime"}.
\item \code{fieldName} - character - The name of the column.
\item \code{index} - integer - The column number.
\item \code{isReferenced} - logical - If \code{TRUE}, then the column is referenced in the worksheet.
}}
}}

\seealso{
Data table schema objects are obtained via
\code{\link[=tableau_worksheet_info]{tableau_worksheet_info()}} and \code{\link[=tableau_datasource_info]{tableau_datasource_info()}}.
}
