% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableauextension.R
\name{ext_options}
\alias{ext_options}
\title{Specify options for \code{tableau_extension}}
\usage{
ext_options(
  config_width = 600,
  config_height = 400,
  prompt_for_config = TRUE,
  use_theme = TRUE,
  ...
)
}
\arguments{
\item{config_width, config_height}{Numeric value specifying the initial width
and height of the extension's configuration dialog (if any). In pixels.}

\item{prompt_for_config}{If the extension provides a configuration dialog
(i.e. \code{\link[=tableau_extension]{tableau_extension()}} is called with a \code{config_ui} argument),
\code{prompt_for_config=TRUE} (the default) means that a newly added extension
should not even attempt to render its \code{ui} and \code{server} in the dashboard;
instead, a message instructing the Tableau user to use the configuration
dialog is displayed instead.

Use \code{prompt_for_config=FALSE} if the extension is able to run even without
initial configuration.}

\item{use_theme}{By default, shinytableau applies a customized version of
Bootstrap 4 CSS to your extension's UI and config UI; the customizations
are intended to complement Tableau's own UI conventions (though we cannot
replicate them exactly due to licensing issues). Pass \code{FALSE} if you want
to omit shinytableau's styles and just use your own.}

\item{...}{Options to pass through to \link[shiny:runApp]{shiny::runApp} (e.g. \code{port},
\code{launch.browser}, \code{host}, \code{quiet}). For local development purposes, it's a
good idea to assign \code{port} to a hardcoded number between 1025 and 49151
that is unique for each extension.}
}
\value{
A list object, suitable for passing to \code{tableau_extension(options = ...)}.
}
\description{
Use this function to modify the configuration options available for your
Tableau extension. Pass the resulting value as the \code{options} argument to
\code{\link[=tableau_extension]{tableau_extension()}}.
}
