% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-accessors.R
\name{reactive_tableau_data}
\alias{reactive_tableau_data}
\title{Construct a reactive expression that reads Tableau data}
\usage{
reactive_tableau_data(
  spec,
  options = list(),
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{spec}{An argument that specifies what specific data should be
retrieved. This can be specified in a number of ways:
\enumerate{
\item The name of a setting, that was set using a value returned from
\code{\link[=choose_data]{choose_data()}}. This is the most common scenario for \code{server}.
\item The object returned from \code{\link[=choose_data]{choose_data()}} can be passed in directly. This
is likely the approach you should take if you want to retrieve data in
\code{config_server} based on unsaved config changes (e.g. to give the user a
live preview of what their \code{choose_data} choices would yield).
\item You can directly create a spec object using one of the helper functions
\code{\link[=spec_summary]{spec_summary()}}, \code{\link[=spec_underlying]{spec_underlying()}}, or \code{\link[=spec_logical]{spec_logical()}}. For cases where
the data is not selected based on \code{\link[=choose_data]{choose_data()}} at all, but
programmatically determined or hardcoded. (This should not be common.)
}}

\item{options}{A named list of options:
\itemize{
\item \code{ignoreAliases} - Do not use aliases specified in the data source in
Tableau. Default is \code{FALSE}.
\item \code{ignoreSelection} - If \code{FALSE} (the default), only return data for the
currently selected marks. Does not apply for datasource tables, only
summary and underlying.
\item \code{includeAllColumns} - Return all the columns for the table. Default is
\code{FALSE}. Does not apply for datasource and summary tables, only underlying.
\item \code{maxRows} - The maximum number of rows to return. \strong{Tableau will not,
under any circumstances, return more than 10,000 rows for datasource and
underlying tables.} This option is ignored for summary tables.
\item \code{columnsToInclude} - Character vector of columns that should be included;
leaving this option unspecified means all columns should be returned. Does
not apply for summary and underlying, only datasource.
}}

\item{session}{The Shiny \code{session} object. (You should probably just use the
default.)}
}
\description{
This function is used to read data from Tableau. Because of the many levels
of indirection involved in actually physically reading data from Tableau,
using this function is significantly more involved than, say, a simple
\code{\link[=read.csv]{read.csv()}}. See the Details section for a more detailed introduction.
}
\details{
There are two complicating factors when reading data from Tableau; the first
is how to tell shinytableau what specific data table you want to access, and
the second is actually accessing the data from R.
\subsection{Specifying a data table}{

As of today, Tableau's Extension API requires you to access any data tables
via one of the worksheets on the same dashboard.
}
}
\examples{
data_spec_x <- choose_data("x", iv = iv)
data_x <- reactive_tableau_data(data_spec_x)

}
