% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-accessors.R
\name{reactive_tableau_schema}
\alias{reactive_tableau_schema}
\title{Construct a reactive expression that reads Tableau data schema}
\usage{
reactive_tableau_schema(spec, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{spec}{See \code{\link[=reactive_tableau_data]{reactive_tableau_data()}}.}

\item{session}{The Shiny \code{session} object. (You should probably just use the
default.)}
}
\value{
A named list, as described in the \link{DataTableSchema} topic.
}
\description{
Creates a reactive expression that returns schema data for the specified
Tableau data table, including the names and data types of columns. Basically,
this is a convenience wrapper that takes a \code{spec} object in any of its
various forms, invokes either \code{\link[=tableau_worksheet_info]{tableau_worksheet_info()}} or
\code{\link[=tableau_datasource_info]{tableau_datasource_info()}} as appropriate, and extracts the specific
sub-object that matches \code{spec}.
}
