% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{restore_inputs}
\alias{restore_inputs}
\title{Prepopulate Shiny inputs}
\usage{
restore_inputs(..., session. = getDefaultReactiveDomain())
}
\arguments{
\item{...}{Named arguments, where the names are input names (don't forget to
use namespaces if you're using modules!) and the values are default values
for the corresponding inputs.}

\item{session.}{The current Shiny session.}
}
\value{
\code{session.}, invisibly.
}
\description{
Call this before your Shiny inputs are created, and when they are created,
they will prepopulate themselves to the values contained herein. For example,
if you call \code{restore_inputs(x = 3)}, and then subsequently a
\code{numericInput("x", "x", 5)} is created, the created input will default to \code{3}
instead of \code{5}. Note that each call to \code{restore_inputs} completely replaces
any bookmarkable state, or previous calls to \code{restore_inputs}.
}
\details{
\code{restore_inputs} is designed to be called from (the beginning of) a Shiny
server function, and as such, can only affect inputs created on the server
side (generally using \code{renderUI} or \code{insertUI}).
}
