% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{shinytableau_theme}
\alias{shinytableau_theme}
\title{Bootstrap 4 theme for shinytableau}
\usage{
shinytableau_theme()
}
\description{
Use the return value for \code{shinytableau_theme()} as the \code{theme} argument for
\code{\link[shiny:fillPage]{shiny::fillPage()}}, \code{\link[shiny:fluidPage]{shiny::fluidPage()}}, etc. to add some CSS rules that
make some Shiny controls look more like Tableau. This is only necessary for
\code{ui} objects, not \code{config_ui}, as the latter includes this automatically.
}
