% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-accessors.R
\name{spec_summary}
\alias{spec_summary}
\alias{spec_underlying}
\alias{spec_datasource}
\title{Create data spec objects programmatically}
\usage{
spec_summary(worksheet = 1L, session = shiny::getDefaultReactiveDomain())

spec_underlying(
  worksheet = 1L,
  underlyingTableId = 1L,
  session = shiny::getDefaultReactiveDomain()
)

spec_datasource(
  worksheet = 1L,
  dataSourceId = 1L,
  logicalTableId = 1L,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{worksheet}{The name (as character vector) or number (as integer) of the
worksheet. If a number is given, it will immediately be resolved to a
worksheet name.}

\item{session}{The Shiny \code{session} object. (You should probably just use the
default.)}

\item{underlyingTableId, dataSourceId, logicalTableId}{The id (as character
vector) or number (as integer) of the specific underlying table/data
source/logical table to read. If a number is given, it will immediately be
resolved to an id.}
}
\value{
A spec object, suitable for the \code{spec} argument to
\code{\link[=reactive_tableau_data]{reactive_tableau_data()}} or persisting via
\code{\link[=update_tableau_settings_async]{update_tableau_settings_async()}}.
}
\description{
A data spec object is a pointer to a specific data table in a Tableau
dashboard. It is analogous to a file path or a URL, except instead of a
simple string, it is a structured object consisting of multiple arguments.
The components of each data spec object will vary, depending on the type of
data being requested: summary, underlying, or data source. See the Details
section of \code{\link[=reactive_tableau_data]{reactive_tableau_data()}} for more information.
}
