% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{tableau_close_dialog}
\alias{tableau_close_dialog}
\title{Close Tableau extension's configuration dialog}
\usage{
tableau_close_dialog(payload = "", session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{payload}{Not currently used by shinytableau.}

\item{session}{The Shiny \code{session} object. (You should probably just use the
default.)}
}
\description{
Invoke from within \code{config_server} function to close the configuration
dialog. Most extensions should not need to call this, as shinytableau will
automatically call it when the configuration dialog's OK or Cancel buttons
are pressed.
}
