% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-accessors.R
\name{tableau_datasources}
\alias{tableau_datasources}
\alias{tableau_datasource_info}
\title{Get info about available data sources in this Tableau dashboard}
\usage{
tableau_datasources(session = shiny::getDefaultReactiveDomain())

tableau_datasource_info(id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{session}{The Shiny \code{session} object. (You should probably just use the
default.)}

\item{id}{A data source ID, as returned by \code{tableau_datasources()} or
\code{\link[=tableau_worksheet_info]{tableau_worksheet_info()}}\verb{$dataSourceIds}.}
}
\value{
\code{tableau_datasources()} returns a character vector whose elements are
data source IDs.

\code{tableau_datasource_info()} returns the metadata and schema for a specific
data source. Note that an extension instance can only access data sources
that are actually used by worksheets in the same dashboard. The return
value is a named list that contains the following fields:
\itemize{
\item \strong{\code{id}} - Unique ID for this data source.
\item \strong{\code{fields}} - data frame where each row is one of the fields in the data source, and these columns:
\itemize{
\item \code{aggregation} - character - The type of aggregation used for this field. Possible values listed \href{https://tableau.github.io/extensions-api/docs/enums/tableau.fieldaggregationtype.html}{here}, e.g. \code{"attr"}, \code{"avg"}, \code{"count"}, ...
\item \code{id} - character - The field id.
\item \code{name} - character - The caption for this field.
\item \code{description} - character - User description of the field, or \code{""} if there is none.
\item \code{role} - character - \code{"dimension"}, \code{"measure"}, or \code{"unknown"}.
\item \code{isCalculatedField} - logical - Whether the field is a table calculation.
\item \code{isCombinedField} - logical - Whether the field is a combination of multiple fields.
\item \code{isGenerated} - logical - Whether this field is generated by Tableau. Tableau generates a number of fields for a data source, such as Number of Records, or Measure Values. This property can be used to distinguish between those fields and fields that come from the underlying data connection, or were created by a user.
\item \code{isHidden} - logical - Whether this field is hidden.
}
\item \strong{\code{isExtract}} - \code{TRUE} if this data source is an extract, \code{FALSE} otherwise.
\item \strong{\code{name}} - The user friendly name of the data source as seen in the UI.
\item \strong{\code{extractUpdateTime}} - A \link{POSIXlt} indicating the time of extract, or \code{NULL} if the data source is live.
\item \strong{\code{logicalTables}} - An unnamed list; each element is a \link[=DataTableSchema]{data table schema}.
}
}
\description{
For advanced use only; most shinytableau extensions should use the
\code{\link[=choose_data]{choose_data()}} module to allow the user to specify a data source.
}
