% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{tableau_is_embedded}
\alias{tableau_is_embedded}
\title{Determine whether a standalone app is running in Tableau}
\usage{
tableau_is_embedded()
}
\value{
\code{TRUE} if running in Tableau, \code{FALSE} if not or if it cannot be
determined.
}
\description{
For extensions that use \code{\link[=ext_options]{ext_options(standalone = TRUE)}}, the
\code{ui} and \code{server} objects are invoked in both standalone mode (outside of a
Tableau dashboard) and embedded mode (inside a dashboard). The
\code{tableau_is_embedded()} function can be used to distinguish between the two
cases, in case you want to vary the UI elements and/or behavior for each.
}
