% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{tableau_select_marks_by_brush_async}
\alias{tableau_select_marks_by_brush_async}
\title{Update a worksheet's selection using a \code{plotOutput} ggplot2 brush}
\usage{
tableau_select_marks_by_brush_async(
  worksheet,
  brush,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{worksheet}{Single-element character vector naming the worksheet whose
selection we want to set. If your extension's configuration dialog uses
\code{\link[=choose_data]{choose_data()}} to allow the user to choose data, then use the
\code{tableau_setting(YOUR_SETTING_ID)$worksheet} slot to access the worksheet
name.}

\item{brush}{\code{input$YOUR_BRUSH_ID}, where \code{YOUR_BRUSH_ID} is the identifier
passed as \code{plotOutput}'s \code{brush} argument (or passed to
\code{\link[shiny:brushOpts]{shiny::brushOpts()}}, if you did it that way).}

\item{session}{The Shiny \code{session} object. (You should probably just use the
default.)}
}
\value{
A \link[promises:promise]{promises::promise} object. The promise object will not resolve to
a useful value, but you can use it to handle errors.
}
\description{
Shiny has its own built-in interactive plots feature (see
\href{https://shiny.rstudio.com/articles/plot-interaction.html}{this} and
\href{https://shiny.rstudio.com/articles/selecting-rows-of-data.html}{this} to get
started). The \code{tableau_select_marks_by_brush_async} function lets you take a
brush input from a \code{\link[shiny:plotOutput]{shiny::plotOutput()}} and, if anything is selected, use it
to drive the selection of a Tableau worksheet within the same dashboard. Note
that because Tableau's selection model operates on marks, not rows in the
underlying data, \strong{the x and y dimensions in the originating plot must be
represented in the summary data of the target worksheet.}
}
\details{
Currently this function only works with plots based on ggplot2, not base
graphics plots, because only ggplot2 plots supply the metadata that we need.
Future releases of Tableau may make it possible to provide this feature to
base plots as well.
}
