% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-accessors.R
\name{tableau_worksheets}
\alias{tableau_worksheets}
\alias{tableau_worksheet_info}
\title{Get info about available worksheets in this Tableau dashboard}
\usage{
tableau_worksheets(session = shiny::getDefaultReactiveDomain())

tableau_worksheet_info(name, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{session}{The Shiny \code{session} object. (You should probably just use the
default.)}

\item{name}{A worksheet name, as returned by \code{tableau_worksheets()}.}
}
\value{
\code{tableau_worksheets()} returns a character vector whose elements are
worksheet names. Note that only worksheets that are included on the same
dashboard will be listed, and these are the only worksheets we can access.

\code{tableau_worksheet_info()} returns metadata for a specific worksheet. The
return value is a named list that contains the following fields:
\itemize{
\item \strong{\code{name}} - The name of the worksheet.
\item \strong{\code{summary}} - The \link[=DataTableSchema]{data table schema object} for the
worksheet's summary-level data table.
\item \strong{\code{dataSourceIds}} - Character vector of data source IDs used by this
worksheet. See \code{\link[=tableau_datasource_info]{tableau_datasource_info()}}.
\item \strong{\code{underlyingTables}} - Unnamed list, each element is a \link[=DataTableSchema]{data table schema object} of one of the worksheet's underlying
data tables.
}
}
\description{
For advanced use only; most shinytableau extensions should use the
\code{\link[=choose_data]{choose_data()}} module to allow the user to specify a worksheet.
}
