% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{tableauChangeParameter}
\alias{tableauChangeParameter}
\title{Change a parameter in a Tableau workbook}
\usage{
tableauChangeParameter(id, parameter, value)
}
\arguments{
\item{id}{the Tableau object id}

\item{parameter}{the name of the parameter to change}

\item{value}{the new value for the parameter}
}
\value{
The Tableau object id is returned for use in \%>\% chains
}
\description{
Makes an asynchronous call to change a parameter across the entire Tableau workbook.
The value should be the same data type as the parameter and within the allowable range of
values. It also needs to be the aliased value and not the raw value.
}

