% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{tableauChangeSheet}
\alias{tableauChangeSheet}
\title{Change the active sheet in a Tableau workbook}
\usage{
tableauChangeSheet(id, worksheet)
}
\arguments{
\item{id}{the Tableau object id}

\item{worksheet}{the name or index of worksheet to switch to}
}
\value{
The Tableau object id is returned for use in \%>\% chains
}
\description{
Makes an asynchronous call to switch to a different sheet in the Tableau workbook.
}

