% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{tableauExportData}
\alias{tableauExportData}
\title{Export Raw Data}
\usage{
tableauExportData(id, worksheet = NULL)
}
\arguments{
\item{id}{the Tableau object id}

\item{worksheet}{the worksheet to download data from, uses the active sheet if \code{NULL}}
}
\value{
The Tableau object id is returned for use in \%>\% chains
}
\description{
Export Raw Data
}

