% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{tableauExportPDF}
\alias{tableauExportPDF}
\title{Export a PDF}
\usage{
tableauExportPDF(id)
}
\arguments{
\item{id}{the Tableau object id}
}
\value{
The Tableau object id is returned for use in \%>\% chains
}
\description{
Export a PDF
}

