% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{tableauExportPNG}
\alias{tableauExportPNG}
\title{Export a PNG}
\usage{
tableauExportPNG(id)
}
\arguments{
\item{id}{the Tableau object id}
}
\value{
The Tableau object id is returned for use in \%>\% chains
}
\description{
Export a PNG
}

