% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{tableauFilter}
\alias{tableauFilter}
\title{Apply a categorical filter to a Tableau view}
\usage{
tableauFilter(id, field, value, type = "REPLACE", exclude = FALSE)
}
\arguments{
\item{id}{the Tableau object id}

\item{field}{the field to filter on}

\item{value}{the value or array of values to filter for}

\item{type}{how the criteria should be applied}

\item{exclude}{a logical value (\code{TRUE} or \code{FALSE}).  If \code{TRUE}
the filter will be applied in exclude mode.}
}
\value{
The Tableau object id is returned for use in \%>\% chains
}
\description{
Makes an asynchronous call to apply a categorical filter to the current Tableau view. Date filters should
only be applied using \code{\link{tableauRangeFilter}}. If the current view is a dashboard, all worksheets
within the dashboard will be filtered. This function works similarly to \code{\link{tableauSelect}}.
}
\section{Filter types}{


The Tableau API supports four filter types:

\describe{
  \item{\code{ALL}}{Adds all values to the filter. Equivalent to checking the (All) value in a quick filter.}

  \item{\code{REPLACE}}{Replaces the current filter values with new ones specified in the call.}

  \item{\code{ADD}}{Adds the filter values as specified in the call to the current filter values.
  Equivalent to checking a value in a quick filter.}

  \item{\code{REMOVE}}{Removes the filter values as specified in the call from the current filter values.
  Equivalent to unchecking a value in a quick filter.}
}
}

