% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{tableauRefresh}
\alias{tableauRefresh}
\title{Refresh the underlying data for a Tableau workbook}
\usage{
tableauRefresh(id)
}
\arguments{
\item{id}{the Tableau object id}
}
\value{
The Tableau object id is returned for use in \%>\% chains
}
\description{
Refresh the underlying data for a Tableau workbook
}

