% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{tableauRevert}
\alias{tableauRevert}
\title{Revert a Tableau workbook to the original settings}
\usage{
tableauRevert(id)
}
\arguments{
\item{id}{the Tableau object id}
}
\value{
The Tableau object id is returned for use in \%>\% chains
}
\description{
Revert a Tableau workbook to the original settings
}

