% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{tableauSelect}
\alias{tableauSelect}
\title{Select data in a Tableau worksheet}
\usage{
tableauSelect(id, field, value, worksheet = NULL, type = "REPLACE")
}
\arguments{
\item{id}{the Tableau object id}

\item{field}{the field to filter on}

\item{value}{the value or array of values to filter for}

\item{worksheet}{the name of the worksheet to apply the selection to}

\item{type}{how the criteria should be applied}
}
\value{
The Tableau object id is returned for use in \%>\% chains
}
\description{
Makes an asynchronous call to select data in the current Tableau view. Multiple worksheetes can not be
selected at the same time. If the current view is a dashboard, the \code{worksheet} argument should be
specified, otherwise the first one alphabetically will be selected. The field must already be available in
the worksheet for this to be applied. This function works similarly to \code{\link{tableauFilter}}.
}
\section{Selection types}{


The Tableau API supports three selection types:

\describe{
  \item{\code{REPLACE}}{Replaces the current marks values with new ones specified in the call.}

  \item{\code{ADD}}{Adds the values as specified in the call to the current selection.
  Equivalent to control-clicking in desktop.}

  \item{\code{REMOVE}}{Removes the values as specified in the call from the current selection.
  Equivalent to control-clicking an already selected mark in desktop.}
}
}

