% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trex.R
\name{tableau_manifest}
\alias{tableau_manifest}
\title{Generate a Tableau Extension Manifest through a function}
\usage{
tableau_manifest(
  extension_id = "com.example.extensions.name",
  extension_version = "0.1.0",
  name = "My extension",
  description = NULL,
  extended_description = description,
  author_name = "Your Name",
  author_email = NULL,
  author_organization = NULL,
  website = "https://example.com",
  source_location = NULL,
  icon = system.file("default_icon.png", package = "shinytableau"),
  permissions = c("full data"),
  configure = NULL,
  min_api_version = "1.4"
)
}
\arguments{
\item{extension_id}{The ID value for the Tableau extension.}

\item{extension_version}{The version number.}

\item{name}{The name of the extension.}

\item{description, extended_description}{The description and extended
description for the extension.}

\item{author_name, author_email, author_organization, website}{Details about the
extension author and project website.}

\item{source_location}{The source location.}

\item{icon}{The icon to be used for the extension.}

\item{permissions}{A setting that determines the level of permissions.}

\item{configure}{Configuration settings.}

\item{min_api_version}{The minimum API version.}
}
\description{
Generate a Tableau Extension Manifest through a function
}
\examples{
# Create a Tableau Manifest via the
# `tableau_manifest()` function
tableau_manifest(
  extension_id = "com.example.ggvoronoi",
  extension_version = "1.1.3",
  name = "Voronoi Plot",
  description = "Insert a Voronoi plot using ggvoronoi",
  author_name = "Jane Doe",
  author_email = "jane_doe@example.com",
  website = "https://example.com/tableau/extensions/ggvoronoi"
)

}
