% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.R
\name{yaml_skeleton}
\alias{yaml_skeleton}
\title{Create a skeleton YAML manifest file}
\usage{
yaml_skeleton(filename = "manifest.yml", path = NULL)
}
\arguments{
\item{filename}{The name of the YAML file to create on disk. By default, this
is `"manifest.yml"`.}

\item{path}{An optional path to which the YAML file should be saved (combined
with `filename`).}
}
\description{
Tableau extensions require a manifest and the use of a YAML manifest file
in **shinytableau** allows for a separation of concerns when developing an
extension. With the `yaml_skeleton()` function, we can write a YAML manifest
file to disk; it contains all of the required fields for the extension with
example values. Later in the development lifecycle, the developer of a
**shinytableau** extension can use values relevant to the project.
}
\examples{
# Generate a YAML manifest file in the
# root directory of the project
# yaml_skeleton()

# Generate the YAML file in a subdirectory
# and call the file 'manifest-temp.yml'
# yaml_skeleton(
#   filename = "manifest-temp.yml",
#   path = "app"
# )

}
