% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.R
\name{yaml_skeleton}
\alias{yaml_skeleton}
\title{Create a skeleton YAML manifest file}
\usage{
yaml_skeleton(filename = "manifest.yml")
}
\arguments{
\item{filename}{A filename for the YAML manifest file. Can be combined with a
relative path. The default, and recommended, filename is `"manifest.yml"`.}
}
\description{
Tableau extensions require a manifest and the use of a YAML manifest file in
**shinytableau** allows for a separation of concerns when developing an
extension. With the `yaml_skeleton()` function, we can write a YAML manifest
file to disk; it contains all of the required fields for the extension with
example values. Later in the development lifecycle, the developer of a
**shinytableau** extension can use values relevant to the project.
}
\examples{
# Generate a YAML manifest file in the
# root directory of the project
# yaml_skeleton()

# Generate the skeleton YAML manifest
# file (called `manifest.yml`) in the
# `app` subdirectory
# yaml_skeleton(filename = "app/manifest.yml")

}
